forge fmt

+++ contracts/deployer/DeterministicDeployFactory.sol
@@ -4,13 +4,11 @@ pragma solidity ^0.8.9;
 contract DeterministicDeployFactory {
     event Deploy(address addr);
 
-    function deploy(bytes memory bytecode, uint _salt) external {
+    function deploy(bytes memory bytecode, uint256 _salt) external {
         address addr;
         assembly {
             addr := create2(0, add(bytecode, 0x20), mload(bytecode), _salt)
-            if iszero(extcodesize(addr)) {
-                revert(0, 0)
-            }
+            if iszero(extcodesize(addr)) { revert(0, 0) }
         }
 
         emit Deploy(addr);
